修改entrypoint.sh

+++ entrypoint.sh
@@ -51,37 +51,40 @@ fi
 CRUDINI='/usr/bin/crudini'
 
 CONNECTION=mysql://nova:$NOVA_DBPASS@$NOVA_DB/nova
+if [ ! -f /etc/nova/.complete ];then
+    cp -rp /nova/* /etc/nova
 
-chown glance:glance /var/log/glance/
+    chown nova:nova /var/log/nova/
     
-$CRUDINI --set /etc/nova/nova.conf database connection $CONNECTION
+    $CRUDINI --set /etc/nova/nova.conf database connection $CONNECTION
 
-$CRUDINI --set /etc/nova/nova.conf DEFAULT rpc_backend rabbit
+    $CRUDINI --set /etc/nova/nova.conf DEFAULT rpc_backend rabbit
 
-$CRUDINI --set /etc/nova/nova.conf oslo_messaging_rabbit rabbit_host $RABBIT_HOST
-$CRUDINI --set /etc/nova/nova.conf oslo_messaging_rabbit rabbit_userid $RABBIT_USERID
-$CRUDINI --set /etc/nova/nova.conf oslo_messaging_rabbit rabbit_password $RABBIT_PASSWORD
+    $CRUDINI --set /etc/nova/nova.conf oslo_messaging_rabbit rabbit_host $RABBIT_HOST
+    $CRUDINI --set /etc/nova/nova.conf oslo_messaging_rabbit rabbit_userid $RABBIT_USERID
+    $CRUDINI --set /etc/nova/nova.conf oslo_messaging_rabbit rabbit_password $RABBIT_PASSWORD
 
-$CRUDINI --set /etc/nova/nova.conf DEFAULT auth_strategy keystone
+    $CRUDINI --set /etc/nova/nova.conf DEFAULT auth_strategy keystone
 
-$CRUDINI --del /etc/nova/nova.conf keystone_authtoken
+    $CRUDINI --del /etc/nova/nova.conf keystone_authtoken
 
-$CRUDINI --set /etc/nova/nova.conf keystone_authtoken auth_uri http://$KEYSTONE_ENDPOINT:5000
-$CRUDINI --set /etc/nova/nova.conf keystone_authtoken auth_url http://$KEYSTONE_ENDPOINT:35357
-$CRUDINI --set /etc/nova/nova.conf keystone_authtoken auth_url auth_plugin password
-$CRUDINI --set /etc/nova/nova.conf keystone_authtoken project_domain_id default
-$CRUDINI --set /etc/nova/nova.conf keystone_authtoken user_domain_id default
-$CRUDINI --set /etc/nova/nova.conf keystone_authtoken project_name service
-$CRUDINI --set /etc/nova/nova.conf keystone_authtoken username nova
-$CRUDINI --set /etc/nova/nova.conf keystone_authtoken password $NOVA_PASS
+    $CRUDINI --set /etc/nova/nova.conf keystone_authtoken auth_uri http://$KEYSTONE_ENDPOINT:5000
+    $CRUDINI --set /etc/nova/nova.conf keystone_authtoken auth_url http://$KEYSTONE_ENDPOINT:35357
+    $CRUDINI --set /etc/nova/nova.conf keystone_authtoken auth_url auth_plugin password
+    $CRUDINI --set /etc/nova/nova.conf keystone_authtoken project_domain_id default
+    $CRUDINI --set /etc/nova/nova.conf keystone_authtoken user_domain_id default
+    $CRUDINI --set /etc/nova/nova.conf keystone_authtoken project_name service
+    $CRUDINI --set /etc/nova/nova.conf keystone_authtoken username nova
+    $CRUDINI --set /etc/nova/nova.conf keystone_authtoken password $NOVA_PASS
 
-$CRUDINI --set /etc/nova/nova.conf DEFAULT my_ip $MY_IP
+    $CRUDINI --set /etc/nova/nova.conf DEFAULT my_ip $MY_IP
 
-$CRUDINI --set /etc/nova/nova.conf glance host $GLANCE_ENDPOINT
+    $CRUDINI --set /etc/nova/nova.conf glance host $GLANCE_ENDPOINT
 
-$CRUDINI --set /etc/nova/nova.conf oslo_concurrency lock_path /var/lib/nova/tmp
+    $CRUDINI --set /etc/nova/nova.conf oslo_concurrency lock_path /var/lib/nova/tmp
 
-
+    touch /etc/glance/.complete
+fi
 # 同步数据库
 echo 'select * from instances limit 1;' | mysql -h$NOVA_DB  -unova -p$NOVA_DBPASS nova
 if [ $? != 0 ];then
